Move reference and numpy simulators to simulation module
