+++ pyquil/simulation/__init__.py
@@ -1,0 +1,11 @@
+from pyquil.simulation._numpy import (
+    NumpyWavefunctionSimulator,
+    get_measure_probabilities,
+    targeted_einsum,
+    targeted_tensordot,
+)
+from pyquil.simulation._reference import (
+    ReferenceDensitySimulator,
+    ReferenceWavefunctionSimulator,
+    zero_state_matrix,
+)
