+++ pyquil/simulation/tests/test_reference_density.py
@@ -11,7 +11,7 @@ from pyquil.gates import CNOT, H, I, MEASURE, PHASE, RX, RY, RZ, X
 from pyquil.operator_estimation import measure_observables
 from pyquil.paulis import sI, sX, sY, sZ
 from pyquil.pyqvm import PyQVM
-from pyquil.reference_simulator import ReferenceDensitySimulator, _is_valid_quantum_state
+from pyquil.simulation._reference import ReferenceDensitySimulator, _is_valid_quantum_state
 from pyquil.simulation.tools import lifted_gate_matrix
 from pyquil.tests.utils import DummyCompiler
