+++ pyquil/unitary_tools.py
@@ -1,5 +1,5 @@
 ##############################################################################
-# Copyright 2016-2018 Rigetti Computing
+# Copyright 2016-2019 Rigetti Computing
 #
 #    Licensed under the Apache License, Version 2.0 (the "License");
 #    you may not use this file except in compliance with the License.
